Move UserAgentBuilder to transport/http
